fix:deploy error file name
