+++ src/app/pages/adminDashboard/page.tsx
@@ -1,9 +1,9 @@
 "use client"
 
 import { useState } from 'react';
-import { useProjects } from '@/app/hooks/useProjects/page';
+import { useProjects } from '@/app/hooks/useProjects';
 import { useAuth } from '@/app/hooks/useAuth';
-import { useProjectForm } from '@/app/hooks/useProjectForm/page';
+import { useProjectForm } from '@/app/hooks/useProjectForm';
 import { useProjectUtils } from '@/app/hooks/useProjectUtils/page';
 import { useEditProject, type Project } from '@/app/hooks/useEditProject';
 import { Plus, Eye, Edit, Trash2, Search, Filter, Calendar, DollarSign, Users, LogOut, X } from 'lucide-react';
